examples/requirements.txt
pyserial
pandas
numpy
matplotlib